+++ site_builder/formatter/pvme_settings.py
@@ -61,6 +61,16 @@ class PVMEChannelData(dict):
     def __init__(self, url="https://raw.githubusercontent.com/pvme/pvme-settings/pvme-discord/channels.json"):
         channels = {channel['id']: {'path': channel['path'], 'name': channel['name']} for channel in
                     github_json_request(url)}
+
+        # remove pvm-help (was used in some testing in #website but gives 404)
+        # todo: remove when pvm-help is removed from channels.txt
+        pvm_help_id = None
+        for id_, settings in channels.items():
+            if settings.get('name') == 'pvm-help':
+                pvm_help_id = id_
+        if pvm_help_id:
+            channels.pop(pvm_help_id)
+
         super().__init__(channels)