+++ docs/about/changelog.md
@@ -7,6 +7,11 @@ Updates for the website.
 
 For bugs or feature requests, message me on Discord (Pleb#0025) or create an issue on [Github](https://github.com/pvme/pvme.github.io).
 
+## November 23, 2024
+
+- Updated various #unknown-channel to show actual channel name
+- youtube.com/live show embed preview correctly
+
 ## June 28, 2024
 - Added "edit this guide" button
 - Change #user -> @user